+++ CHANGELOG.md
@@ -20,6 +20,7 @@ After this we will switch probably to real [Semantic Versioning 2.0.0](http://se
 * Fix 1600: Replicas fragment not set 
 * Fix 1579: Added documentation for clearing doubts related to sidecars
 * Fix 1587: HandlerHub refactored
+* Fix 1579: Add ImageChange triggers to application container when resource fragments are used
 * Fix 1607: Fix YAML aggregation for OpenShift
 * Refactor 1592: Removed Deprecated/Unused Osio specific enrichers.